pull up generate_tests method into common trait + default implementation
